chore: reorder components to layout level
